IMPALA-13524: Calcite planner: support for functions in exprs.test
Specifically, this commit supports translate, the not operator (!),
the null not equals operator and some "not regexp" functions.

A little extra code had to be added for the not regexp functions
because Calcite treats the "not" part as an attribute to the operator.

Also, added support for "is distinct from" and "is not distinct from"
operators which required a convertlet change to handle the conversion
from SqlNodes to RexNodes.

Change-Id: Ib8c5d5719a409a32ddb6946d1a87c77773f20820
Reviewed-on: http://gerrit.cloudera.org:8080/22104
Reviewed-by: Aman Sinha <amsinha@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>