IMPALA-13571: Calcite Planner: Fix join parsing errors.
This commit fixes two parsing errors related to joins.

1) The straight_join hint is now parsed correctly. The hint, however,
will be ignored.  IMPALA-13574 has been filed for this feature.

2) The anti-join and semi-join will no longer throw parse exceptions
but will now throw unsupported exceptions.  The jiras IMPALA-13572
and IMPALA-13573 have been filed for these features.

IMPALA-13708: Throw unsupported message for complex column syntax.

There is special complex column syntax where the column is treated
like a table and looks like a table name when parsed. If we get an
analysis error for an unfound table, we can see if it is actually
a complex column and throw an "unsupported" error instead of giving
a cryptic "table not found" message.

Also made the support for types of tables to be a bit more generic
and check if it's not an HdfsTable rather than only throw an
unsupported exception if it's an Iceberg table.

Change-Id: Icd0f68441c84b090ed2cb45de96ccee1054deef5
Reviewed-on: http://gerrit.cloudera.org:8080/22412
Reviewed-by: Aman Sinha <amsinha@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>