IMPALA-13539: Fix TestCalcitePlanner.test_calcite_frontend for non-HDFS
TestCalcitePlanner.test_calcite_frontend runs an explain plan
and matches the expected output. On HDFS that includes "SCAN HDFS",
but on S3 and other non-HDFS filesystems, that can be "SCAN S3".
Similarly, "HDFS partitions" can be "S3 partitions". This changes
the expected output to make it insensitive to these differences.

Testing:
 - Ran test case locally and on S3

Change-Id: I7ddf919558a7ad622d7571283ad8a2e069c69f14
Reviewed-on: http://gerrit.cloudera.org:8080/22044
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>