IMPALA-13383: Fix TestIcebergTable.test_create_table_like_table exhaustive test failure
TestIcebergTable.test_create_table_like_table contained 2 tables
that were created in the default Hadoop Catalog location:
'iceberg.catalog_location' = 'ice_hadoop_cat'.
As a result of this, the tables were created under the path defined in
hive-site.xml.py and not in the unique database directory.
They both had a constant string as the table identifier that was
concatenated to the catalog location to form the table's path.
When the test was run in parallel, it failed with AlreadyExistsException
because all tests created the table with the same name in the same
directory.
Fix: Added unique database to the table identifier so that it became
unique and the tables are stored in separate directories.

Testing:
- Ran TestIcebergTable.test_create_table_like_table locally
  in exhaustive mode multiple times before and after the fix.
  Command:
 'tests/run-tests.py --workload_exploration_strategy=functional-query:exhaustive
  query_test/test_iceberg.py::TestIcebergTable::test_create_table_like_table'

Change-Id: I1d70855f37aae06cc2b8e59fb773013e2b3918a8
Reviewed-on: http://gerrit.cloudera.org:8080/22072
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>