IMPALA-13560: Fix test_admission_controller_with_quota_configs test.
The intention of test_admission_controller_with_quota_configs() is to
run a workload with a variety of outcomes in a pool that has Admission
Control User Quotas Configured. The idea was that the User Quotas
configuration would not affect the workload that is run by
run_admission_test(). The configuration for 'queueF' limits the number
of concurrent queries that can be run by any user to 30. In the test
there is only one user, and the number of queries that are run is 50,
so there is potential for the User Quotas configuration to affect the
operation of the test. Fix this by bumping the Quota limit to 50.

TESTING

I ran tests in a similar environment to that where failures were
observed. Without the fix I saw a failure, and with the fix there were
no failures. This isn't sufficient to prove this fix is all that is
needed, but the change is safe and isolated.

Change-Id: Ie2cc81a5b95d07154b73d32daf67617c79283ac8
Reviewed-on: http://gerrit.cloudera.org:8080/22096
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>