IMPALA-13560: Second attempt at fixing test_admission_controller_with_quota_configs test.
The test test_admission_controller_with_quota_configs() is designed to
be a similar test to test_admission_controller_with_configs() which
uses the 'queueB' queue. The newer test uses a newly added
queue 'queueF'. Because Admission Control configuration is split across
two files, and because of user stupidity, the queue timeout
configuration for 'queueB' was not copied when the new test was
created. This causes queued queries to be timed out while waiting for
admission, which confuses the test.

Set pool-queue-timeout-ms.root to 600000 for queueF in
llama-site-test2.xml.

Change-Id: I1378cd4e42ed5629b92b1c16dd17d4d16ec4a19d
Reviewed-on: http://gerrit.cloudera.org:8080/22126
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>