IMPALA-13596: Add warnings and exceptions to reading of fair-scheduler file.
The fair-scheduler file contains part of the configuration for Admission
Control. This change adds some better error handling to the parsing of
this file. Where it is safe to do so, new exceptions are thrown; this
will cause Impala to refuse to start. This is consistent with other
serious configuration errors. Where new exceptions might cause problems
with existing configurations, or for less dangerous faults, new warnings
are written to the server log.

For the recently added User Quota configuration (IMPALA-12345) throw an
exception when a duplicate snippet of configuration is found.

New warning log messages are added for these cases:
- when a user quota at the leaf level is completely ignored because of
  a user quota at the root level
- when there is no user ACL on a leaf level queue. This prevents any
  queries from being submitted to the queue.

Change-Id: Idcd50442ce16e7c4346c6da1624216d694f6f44d
Reviewed-on: http://gerrit.cloudera.org:8080/22209
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>