IMPALA-13566: Expose query cancellation status to UDFs
Previously, the evaluation of a UDF was not interruptible. When the
impalad was evaluating a UDF on a batch, the client had to wait
for the end of processing the batch before it could get the results or
cancel the query. Such a period could be long if the UDF takes a long
time to run, e.g. the sleep() UDF.

This patch tries to mitigate this issue by exposing the query
cancellation status in RuntimeState to the UDF so that the UDF can check
the status by itself and return early. This can significantly reduce the
waiting time when cancelling a long-running query.

As an example, this patch makes the sleep() UDF interruptible by
checking the query cancellation status inside the UDF every 200ms,
controlled by variable SLEEP_UNINTERRUPTIBLE_INTERVAL_MS. This means
that we only need to wait about 200ms to cancel the query.

Testing:
- Added a new test case in tests/query_test/test_cancellation.py to
  ensure that we can interrupt sleep().
- Added a new connection class called MinimalHS2Connection in
  tests/common/impala_connection.py to support manipulating one
  operation from multiple connections for the new test case.

Change-Id: I9430167f7e46bbdf66153abb4645541cd8cf0142
Reviewed-on: http://gerrit.cloudera.org:8080/22280
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>