IMPALA-11979: Add 'scheduling_seed' to customize consistent scheduling behavior
This adds a startup parameter --scheduling_seed which is a string
identifier of an executor within an executor group. It should
be unique within an executor group, but it could be reused
across executors groups on the same system. This is used for
scan range scheduling for remote filesystems, so this can be
used to make the scheduling deterministic across multiple
executor groups or when an executor group gets restarted on
machines with different IP addresses.

For example, the 3rd executor in an executor group of size 8
might use "executor_3_of_8" for its scheduling seed. If there
are multiple executors groups of size 8, the 3rd in each can
use that scheduling seed.

Testing:
 - Ran core job

Change-Id: Ie01c7d119cc88766082dbfca3ff685354d01f71f
Reviewed-on: http://gerrit.cloudera.org:8080/22214
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>