IMPALA-13608: Fix TestIcebergTable.test_catalogs exhaustive test failure
TestIcebergTable.test_catalogs contained 2 tables
that were created in the default Hadoop Catalog location:
'iceberg.catalog_location' = 'ice_hadoop_cat'.
As a result of this, the tables were created under the path defined in
hive-site.xml.py and not in the unique database directory.
They both had a constant string as the table identifier that was
concatenated to the catalog location to form the table's path.
When the test was run in parallel, it failed with AlreadyExistsException
because all tests created the table with the same name in the same
directory.
Fix: Added unique database to the table identifier so that it became
unique and the tables are stored in separate directories.

Testing:
- Ran TestIcebergTable.test_catalogs locally
  in exhaustive mode multiple times before and after the fix.
  Command:
 'tests/run-tests.py --workload_exploration_strategy=functional-query:exhaustive
  query_test/test_iceberg.py::TestIcebergTable::test_catalogs'

Change-Id: If193945f11d450000523a194acd56936a364a9e1
Reviewed-on: http://gerrit.cloudera.org:8080/22218
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>