IMPALA-13630: Avoid holding global lock when closing KrpcDataStreamRecvr
KrpcDataStreamRecvr::CancelStream() may take some time as it:
- may need to wait for SenderQueue's lock
- needs to respond all deferred RPCs

While responding RPCs is asynchronous, it still involves serialization
of the response and a potential system call to schedule the response to
the KRPC reactor thread.

The change is safe as all KrpcDataStreamRecvr::SenderQueue functions
check cancallation and return fast if it is already cancelled. Calling
CancelStream() twice (once in DeregisterRecvr(), once in Cancel()) is
also harmless.

Change-Id: Ie110722f144400132de1e23813e260dab1de77e7
Reviewed-on: http://gerrit.cloudera.org:8080/22251
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>